Refactor SuperGrid, move style calculations to utils
